Delete src/test.py:Zone.Identifier
I think these :Zone.Identifier files were created automatically when you created the other files (I think it is done by the operating system). Are you on Linux, Windows or Linux-virtual machine on Windows?

Problem is that the names of these files is not accepted in Windows so it creates problems when cloning the repo and pulling. I think there is a setting that stops them from being created.